requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pytest==7.4.3
gunicorn==21.2.0
waitress==2.1.2 